+++ packages/core/interface/connection.ts
@@ -68,14 +68,15 @@ export interface ConnectionInterface {
    */
   previewLimit?: number;
   /**
-   * Dialect driver options. See more on https://github.com/mtxr/vscode-sqltools/wiki/connection-driver-options
+   * MSSQL specific driver options. See https://mtxr.gitbook.io/vscode-sqltools/connections/mssql#1-1-specific-options
+   * @deprecated replaced by mssqlOptions
    * @type {any}
    * @memberof ConnectionInterface
    */
-  dialectOptions?: { encrypt?: boolean };
+  mssqlOptions?: { encrypt?: boolean };
 
   /**
-   * MySQL specific options
+   * MySQL specific driver options
    * @type {any}
    * @memberof ConnectionInterface
    */
@@ -84,7 +85,7 @@ export interface ConnectionInterface {
   }
 
   /**
-   * PostgreSQL specific options. See https://mtxr.gitbook.io/vscode-sqltools/connections/postgresql#1-1-specific-options
+   * PostgreSQL specific driver options. See https://mtxr.gitbook.io/vscode-sqltools/connections/postgresql#1-1-specific-options
    * @type {any}
    * @memberof ConnectionInterface
    */