+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * **NEW** 🎉
   * Added option to open results rows/values in editor. Issue #140.
   * Added SSL support for postgres viw `pgOptions` connection setting. Issue #141
+  * Deprecating `dialectOptions` in favor of `[dialect]Options` settings.
 
 * **Fixes**